requirements.txt
tensorflow==2.16.2
streamlit==1.24.0
gdown==4.7.0
numpy==1.24.2
Pillow==9.5.0
setuptools
